+++ test/tcf/services/helpers/TcfRepositoryMock.js
@@ -1,4 +1,4 @@
-import {TcfRepository} from '../../../../components/tcf/services/src/infrastructure/Tcf/TcfRepository'
+import {TcfRepository} from '../../../../components/tcf/services/src/infrastructure/tcfx/TcfRepository'
 class TcfRepositoryMock extends TcfRepository {
   constructor({tcfApi}) {
     super({tcfApi})
